requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
email-validator==2.2.0
databases[asyncpg]==0.9.0
psycopg2==2.9.10
python-jose==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.15.2
# Pin bcrypt until passlib supports the latest
bcrypt==4.0.1
# For logging
rich==13.9.4
asgi-correlation-id==4.3.4
python-json-logger==3.3.0